refactor(contacts): tabs at toolbar level with contextual actions, rename to Serviços
